Refactored dataset into smaller components
